Move creating tray icon to App
